pyproject.toml
[project]
name = "tzafon"
version = "0.1.4"
description = "Tzafon Waypoint – browser automation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "twine>=6.1.0",
    "websockets>=15.0.1",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "python-dotenv>=1.1.0",
    "httpx>=0.26.0",
]
license = { text = "MIT"}

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/tzafon"]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
